pyproject.toml
# pyproject.toml
[tool.poetry]
name = "gather-manager"
version = "0.1.0"
description = "Tools for managing Gather.town spaces and containers"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.28.2"
pydantic = "^2.3.0"
typer = "^0.9.0"
rich = "^13.4.2"
python-dotenv = "^1.0.0"
# Tool inventory manager dependencies
pandas = "^1.3.0"
tabulate = "^0.8.9"

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"
pytest-cov = "^4.1.0"
black = "^23.3.0"
isort = "^5.12.0"
mypy = "^1.3.0"
flake8 = "^6.0.0"
pre-commit = "^3.3.2"
pytest-mock = "^3.10.0"
responses = "^0.23.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
gather-manager = "gather_manager.cli.main:app"

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false

[tool.black]
line-length = 88
target-version = ['py39']

[tool.isort]
profile = "black"
multi_line_output = 3
